performance improvements, fixes, cleanup

+++ dan_layer/wallet/storage_sqlite/src/reader.rs
@@ -920,13 +920,13 @@ impl WalletStoreReader for ReadTransaction<'_> {
         use crate::schema::authored_templates;
 
         let total_templates_for_key_index = authored_templates::table
-            .filter(authored_templates::key_index.eq(key_index as i32))
+            .filter(authored_templates::key_index.eq(key_index as i64))
             .count()
             .first::<i64>(self.connection())
             .map_err(|e| WalletStorageError::general("count_authored_templates_fetch_by_key_index", e))?;
 
         let templates = authored_templates::table
-            .filter(authored_templates::key_index.eq(key_index as i32))
+            .filter(authored_templates::key_index.eq(key_index as i64))
             .limit(page_size as i64)
             .offset((page * page_size) as i64)
             .select(AuthoredTemplate::as_select())
